Split parts into separate files
