+++ src/__init__.py
@@ -1,0 +1,5 @@
+"""GearExporter JSON to cmangos pdump SQL importer."""
+
+from src.parser import parse_file
+
+__all__ = ["parse_file"]
