+++ src/items/__init__.py
@@ -1,0 +1,15 @@
+from src.items.assembly import _add_to_itemlists
+from src.items.enchantments import (
+    _build_enchantments,
+)
+from src.items.normalizer import (
+    _default_gems,
+    _normalize_item_fields,
+)
+
+__all__ = [
+    "_add_to_itemlists",
+    "_build_enchantments",
+    "_default_gems",
+    "_normalize_item_fields",
+]
