+++ src/parsers/__init__.py
@@ -1,0 +1,28 @@
+from src.parsers.achievements import _parse_achievements
+from src.parsers.actions import _parse_actions
+from src.parsers.bags import _parse_bag_contents
+from src.parsers.equipment import _parse_equipment
+from src.parsers.factions import _parse_factions
+from src.parsers.glyphs import _parse_glyphs
+from src.parsers.macros import _parse_macros
+from src.parsers.pet import _parse_pet
+from src.parsers.quests import _parse_quests
+from src.parsers.skills import _add_default_skills, _parse_char_skills
+from src.parsers.spells import _parse_spells
+from src.parsers.talents import _parse_talents
+
+__all__ = [
+    "_add_default_skills",
+    "_parse_achievements",
+    "_parse_actions",
+    "_parse_bag_contents",
+    "_parse_char_skills",
+    "_parse_equipment",
+    "_parse_factions",
+    "_parse_glyphs",
+    "_parse_macros",
+    "_parse_pet",
+    "_parse_quests",
+    "_parse_spells",
+    "_parse_talents",
+]
